+++ src/test/kotlin/tornadofx/PreferencesTest.kt
@@ -1,0 +1,28 @@
+package tornadofx
+
+import org.junit.Assert
+import org.junit.Test
+
+class PreferencesTest {
+
+    @Test
+    fun testPreferences() {
+        object : Controller() {
+            init {
+                preferences("test app") {
+                    putBoolean("boolean key", true)
+                    put("string", "value")
+                }
+
+                var bool: Boolean = false
+                var str: String = ""
+                preferences("test app") {
+                    bool = getBoolean("boolean key", false)
+                    str = get("string", "")
+                }
+                Assert.assertEquals(true, bool)
+                Assert.assertEquals("value", str)
+            }
+        }
+    }
+}