+++ src/test/kotlin/tornadofx/PropertiesTest.kt
@@ -43,6 +43,7 @@ class PropertiesTest {
         assertEquals(42, fixture.integer)
         assertEquals(42, fixture.integerDefault)
     }
+
     class TestClass {
         var myProperty: String by singleAssign()
     }
@@ -59,12 +60,14 @@ class PropertiesTest {
         }
         assertTrue(failed)
     }
+
     @Test
     fun succeedAssignment() {
         val instance = TestClass()
         instance.myProperty = "foo"
         instance.myProperty
     }
+
     @Test
     fun failDoubleAssignment() {
 
@@ -79,6 +82,7 @@ class PropertiesTest {
         }
         assertTrue(failed)
     }
+
     @Test
     fun pojoWritableObservable() {
         val person = JavaPerson()
@@ -93,5 +97,9 @@ class PropertiesTest {
 
         Assert.assertEquals(44, person.id)
         Assert.assertEquals("Doe", person.name)
+
+        person.id = 5
+        idObservable.refresh()
+        Assert.assertEquals(5, idObservable.value)
     }
 }